feat: access old terminals
Throught some rework, it is now possible to see the output,
and continue interacting, with a terminal which was once closed.

This can be done programatically, or with the new telescope plugin.

Also, added a small fix where telescope would crash if you
pressed enter when there were no actions.